% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{p_star_cal}
\alias{p_star_cal}
\title{p_star_cal}
\usage{
p_star_cal(p_value)
}
\arguments{
\item{p_value}{the p_value.}
}
\value{
Returns: Signif. codes:  0 "***" 0.001 "**" 0.01 "*" 0.05 "." 0.1 " " 1
}
\description{
Returns * based on p value
}
