% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_liu.R
\name{theme_liu}
\alias{theme_liu}
\title{Title This is the theme for ggplot2 using Linkoping University colors}
\usage{
theme_liu(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base parameters}

\item{base_family}{base parameters}
}
\value{
None
}
\description{
Title This is the theme for ggplot2 using Linkoping University colors
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg, color = cyl)) +
ggplot2::geom_point() + ggplot2::geom_smooth(method = "lm", formula = y~x) +
ggplot2::labs(title = "Regression of MPG on Weight", x = "Weight", y = "Miles per Gallon")
}
